requirements.txt
discord.py==1.3.3
pymongo==3.9.0
requests==2.23.0
dnspython==1.16.0
